# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for customer and subscription management
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface Customer {
  id: number;
  name: string;
  /** @nullable */
  contact?: string | null;
  /** @nullable */
  note?: string | null;
  createdAt: string;
  subscriptionCount: number;
}

export interface CustomerInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  contact?: string | null;
  /** @nullable */
  note?: string | null;
}

export interface CustomerUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  contact?: string | null;
  /** @nullable */
  note?: string | null;
}

export type SubscriptionStatus = typeof SubscriptionStatus[keyof typeof SubscriptionStatus];


export const SubscriptionStatus = {
  active: 'active',
  expiring: 'expiring',
  expired: 'expired',
  archived: 'archived',
} as const;

export interface Subscription {
  id: number;
  customerId: number;
  customerName: string;
  /** @nullable */
  customerContact?: string | null;
  productId: number;
  productName: string;
  /** @nullable */
  sourceAccountId?: number | null;
  /** @nullable */
  sourceAccountEmail?: string | null;
  startDate: string;
  endDate: string;
  /** @nullable */
  price?: number | null;
  status: SubscriptionStatus;
  /** @nullable */
  revokedAt?: string | null;
  daysRemaining: number;
  createdAt: string;
  updatedAt: string;
}

export type CustomerDetail = Customer & {
  subscriptions: Subscription[];
};

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  defaultDurationDays?: number | null;
  /** @nullable */
  defaultPrice?: number | null;
  activeSubscriptions: number;
  sourceAccountCount: number;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  defaultDurationDays?: number | null;
  /** @nullable */
  defaultPrice?: number | null;
}

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  defaultDurationDays?: number | null;
  /** @nullable */
  defaultPrice?: number | null;
}

export interface SourceAccount {
  id: number;
  productId: number;
  productName: string;
  email: string;
  maxSlots: number;
  usedSlots: number;
  availableSlots: number;
  /** @nullable */
  expiresAt?: string | null;
  /** @nullable */
  note?: string | null;
}

export interface SourceAccountInput {
  productId: number;
  /** @minLength 1 */
  email: string;
  /** @minimum 1 */
  maxSlots: number;
  /** @nullable */
  expiresAt?: string | null;
  /** @nullable */
  note?: string | null;
}

export interface SourceAccountUpdate {
  /** @minLength 1 */
  email?: string;
  /** @minimum 1 */
  maxSlots?: number;
  /** @nullable */
  expiresAt?: string | null;
  /** @nullable */
  note?: string | null;
}

export interface SubscriptionInput {
  customerId: number;
  productId: number;
  /** @nullable */
  sourceAccountId?: number | null;
  startDate: string;
  endDate: string;
  /** @nullable */
  price?: number | null;
}

export interface SubscriptionUpdate {
  /** @nullable */
  sourceAccountId?: number | null;
  startDate?: string;
  endDate?: string;
  /** @nullable */
  price?: number | null;
}

export interface RenewalHistory {
  id: number;
  subscriptionId: number;
  renewedAt: string;
  daysAdded: number;
  /** @nullable */
  amountPaid?: number | null;
  gapDays: number;
  /** @nullable */
  note?: string | null;
}

export type SubscriptionDetail = Subscription & {
  history: RenewalHistory[];
};

export interface RenewalInput {
  /** @minimum 1 */
  daysAdded: number;
  /** @nullable */
  amountPaid?: number | null;
  /** @nullable */
  note?: string | null;
}

export interface RevokeInput {
  /** @nullable */
  note?: string | null;
}

export type DashboardSummaryStatusBreakdownItem = {
  status: string;
  count: number;
};

export interface DashboardSummary {
  activeSubscriptions: number;
  expiringSubscriptions: number;
  expiredSubscriptions: number;
  revokedSubscriptions: number;
  totalCustomers: number;
  totalRevenue: number;
  revenueThisMonth: number;
  statusBreakdown: DashboardSummaryStatusBreakdownItem[];
  upcoming: Subscription[];
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  created: 'created',
  renewed: 'renewed',
  revoked: 'revoked',
  expiring: 'expiring',
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  title: string;
  description: string;
  timestamp: string;
  /** @nullable */
  subscriptionId?: number | null;
}

export type IntegrationStatusTelegram = {
  configured: boolean;
  mode: string;
};

export type IntegrationStatusAiRouter = {
  configured: boolean;
  provider: string;
};

export interface IntegrationStatus {
  telegram: IntegrationStatusTelegram;
  aiRouter: IntegrationStatusAiRouter;
}

/**
 * @nullable
 */
export type TelegramUpdateMessage = { [key: string]: unknown } | null;

export interface TelegramUpdate {
  update_id: number;
  /** @nullable */
  message?: TelegramUpdateMessage;
  [key: string]: unknown;
 }

export interface TelegramResult {
  ok: boolean;
  /** @nullable */
  reply: string | null;
}

export interface ReminderRunResponse {
  sent: number;
  skipped: number;
  configured: boolean;
}

/**
 * Bad request
 */
export type BadRequestResponse = Error;

/**
 * Resource not found
 */
export type NotFoundResponse = Error;

export type GetDashboardActivityParams = {
/**
 * @minimum 1
 * @maximum 30
 */
limit?: number;
};

export type ListCustomersParams = {
search?: string;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};

export type ListSourceAccountsParams = {
productId?: number;
search?: string;
};

export type ListSubscriptionsParams = {
status?: ListSubscriptionsStatus;
search?: string;
sort?: ListSubscriptionsSort;
};

export type ListSubscriptionsStatus = typeof ListSubscriptionsStatus[keyof typeof ListSubscriptionsStatus];


export const ListSubscriptionsStatus = {
  all: 'all',
  active: 'active',
  expiring: 'expiring',
  expired: 'expired',
  archived: 'archived',
} as const;

export type ListSubscriptionsSort = typeof ListSubscriptionsSort[keyof typeof ListSubscriptionsSort];


export const ListSubscriptionsSort = {
  end_date: 'end_date',
  created_at: 'created_at',
  customer: 'customer',
} as const;