# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API for customer and subscription management
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 100
      responses:
        "200":
          description: Customer list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "400":
          $ref: "#/components/responses/BadRequest"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDetail"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          $ref: "#/components/responses/NotFound"
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /source-accounts:
    get:
      operationId: listSourceAccounts
      tags: [source-accounts]
      summary: List source accounts
      parameters:
        - name: productId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Source account list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SourceAccount"
    post:
      operationId: createSourceAccount
      tags: [source-accounts]
      summary: Create source account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SourceAccountInput"
      responses:
        "201":
          description: Created source account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceAccount"
  /source-accounts/{id}:
    patch:
      operationId: updateSourceAccount
      tags: [source-accounts]
      summary: Update source account
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SourceAccountUpdate"
      responses:
        "200":
          description: Updated source account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceAccount"
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Customer:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        contact: { type: ["string", "null"] }
        note: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        subscriptionCount: { type: integer }
      required: [id, name, createdAt, subscriptionCount]
    CustomerInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        contact: { type: ["string", "null"] }
        note: { type: ["string", "null"] }
      required: [name]
    CustomerUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        contact: { type: ["string", "null"] }
        note: { type: ["string", "null"] }
    CustomerDetail:
      allOf:
        - $ref: "#/components/schemas/Customer"
        - type: object
          properties:
            subscriptions:
              type: array
              items:
                $ref: "#/components/schemas/Subscription"
          required: [subscriptions]
    Product:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        defaultDurationDays: { type: ["integer", "null"] }
        defaultPrice: { type: ["number", "null"] }
        activeSubscriptions: { type: integer }
        sourceAccountCount: { type: integer }
      required: [id, name, activeSubscriptions, sourceAccountCount]
    ProductInput:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        defaultDurationDays: { type: ["integer", "null"] }
        defaultPrice: { type: ["number", "null"] }
      required: [name]
    ProductUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        defaultDurationDays: { type: ["integer", "null"] }
        defaultPrice: { type: ["number", "null"] }
    SourceAccount:
      type: object
      properties:
        id: { type: integer }
        productId: { type: integer }
        productName: { type: string }
        email: { type: string }
        maxSlots: { type: integer }
        usedSlots: { type: integer }
        availableSlots: { type: integer }
        expiresAt: { type: ["string", "null"], format: date }
        note: { type: ["string", "null"] }
      required: [id, productId, productName, email, maxSlots, usedSlots, availableSlots]
    SourceAccountInput:
      type: object
      properties:
        productId: { type: integer }
        email: { type: string, minLength: 1 }
        maxSlots: { type: integer, minimum: 1 }
        expiresAt: { type: ["string", "null"], format: date }
        note: { type: ["string", "null"] }
      required: [productId, email, maxSlots]
    SourceAccountUpdate:
      type: object
      properties:
        email: { type: string, minLength: 1 }
        maxSlots: { type: integer, minimum: 1 }
        expiresAt: { type: ["string", "null"], format: date }
        note: { type: ["string", "null"] }
    Subscription:
      type: object
      properties:
        id: { type: integer }
        customerId: { type: integer }
        customerName: { type: string }
        customerContact: { type: ["string", "null"] }
        productId: { type: integer }
        productName: { type: string }
        sourceAccountId: { type: ["integer", "null"] }
        sourceAccountEmail: { type: ["string", "null"] }
        startDate: { type: string, format: date }
        endDate: { type: string, format: date }
        price: { type: ["number", "null"] }
        status:
          type: string
          enum: [active, expiring, expired, archived]
        revokedAt: { type: ["string", "null"], format: date-time }
        daysRemaining: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required:
        - id
        - customerId
        - customerName
        - productId
        - productName
        - startDate
        - endDate
        - status
        - daysRemaining
        - createdAt
        - updatedAt
...