# lib/api-zod/src/generated/types/customerUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for customer and subscription management
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  contact?: string | null;
  /** @nullable */
  note?: string | null;
}