# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for customer and subscription management
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  defaultDurationDays?: number | null;
  /** @nullable */
  defaultPrice?: number | null;
  activeSubscriptions: number;
  sourceAccountCount: number;
}