# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for customer and subscription management
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  defaultDurationDays?: number | null;
  /** @nullable */
  defaultPrice?: number | null;
}