# lib/api-zod/src/generated/types/sourceAccount.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for customer and subscription management
 * OpenAPI spec version: 0.1.0
 */

export interface SourceAccount {
  id: number;
  productId: number;
  productName: string;
  email: string;
  maxSlots: number;
  usedSlots: number;
  availableSlots: number;
  /** @nullable */
  expiresAt?: Date | null;
  /** @nullable */
  note?: string | null;
}